add new output file with report summary
new report summary tracks counts of each links from each section of
the report. this can be shared with others to get a quick glance of if
something big changed since the last report. new file name defaults to
linkto_summary.txt and can be changed in the config file under
report->summary.